LayoutTests/fast/scrolling/latching/overflow-hidden-on-one-axis-expected.txt

Tests that scroll events of type 'none' work properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroller.scrollLeft is 0
PASS scroller.scrollTop > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
